refactor: update table cell import/export
