+++ packages/super-editor/src/core/super-converter/v3/handlers/w/tc/index.js
@@ -1,0 +1,1 @@
+export * from './tc-translator';
